Split Route and Router files
