+++ airframe-http/src/main/scala/wvlet/airframe/http/Router.scala
@@ -13,15 +13,10 @@
  */
 package wvlet.airframe.http
 
-import wvlet.airframe.codec.PrimitiveCodec.StringCodec
-import wvlet.airframe.codec.{JSONCodec, MessageCodecFactory}
-import wvlet.airframe.json.JSON
-import wvlet.airframe.surface.reflect._
-import wvlet.airframe.surface.{MethodSurface, Surface, Zero}
+import wvlet.airframe.surface.{MethodSurface, Surface}
 import wvlet.log.LogSupport
 
 import scala.language.experimental.macros
-import scala.util.Try
 
 /**
   * Provides mapping from HTTP requests to controller methods (= Route)
@@ -30,29 +25,18 @@ import scala.util.Try
   */
 class Router(val routes: Seq[Route]) {
 
-  def findRoute[A](request: HttpRequest[A]): Option[Route] = {
-    routes
-      .find { r =>
-        r.method == request.method &&
-        checkPath(request.pathComponents, r.pathComponents)
-      }
-  }
-
-  private def checkPath(requestPathComponents: Seq[String], routePathComponents: Seq[String]): Boolean = {
-    if (requestPathComponents.length == routePathComponents.length) {
-      requestPathComponents.zip(routePathComponents).forall {
-        case (requestPathComponent, routePathComponent) =>
-          routePathComponent.startsWith(":") || routePathComponent == requestPathComponent
-      }
-    } else {
-      false
-    }
-  }
-
   /**
     * Add methods annotated with @Endpoint to the routing table
     */
   def add[Controller]: Router = macro RouterMacros.add[Controller]
+
+  def findRoute[Req](request: HttpRequest[Req]): Option[Route] = {
+    routes
+      .find { r =>
+        r.method == request.method &&
+        Router.checkPath(request.pathComponents, r.pathComponents)
+      }
+  }
 }
 
 object Router extends LogSupport {
@@ -82,101 +66,14 @@ object Router extends LogSupport {
     new Router(r.routes ++ newRoutes)
   }
 
-}
-
-case class Route(controllerSurface: Surface, method: HttpMethod, path: String, methodSurface: ReflectMethodSurface)
-    extends LogSupport {
-  require(
-    path.startsWith("/"),
-    s"Invalid route path: ${path}. EndPoint path must start with a slash (/) in ${methodSurface.owner.name}:${methodSurface.name}")
-
-  val pathComponents: IndexedSeq[String] = {
-    path
-      .substring(1)
-      .split("/")
-      .toIndexedSeq
-  }
-
-  def returnTypeSurface: Surface = methodSurface.returnType
-
-  /**
-    * Extracting path parameter values. For example, /user/:id with /user/1 gives { id -> 1 }
-    */
-  private def extractPathParams[A](request: HttpRequest[A]): Map[String, String] = {
-    val pathParams = (for ((elem, actual) <- pathComponents.zip(request.pathComponents) if elem.startsWith(":")) yield {
-      elem.substring(1) -> actual
-    }).toMap[String, String]
-    pathParams
-  }
-
-  /**
-    * Find a corresponding controller and call the matching methods
-    *
-    * @param request
-    * @return
-    */
-  def buildControllerMethodArgs[A](controller: Any, request: HttpRequest[A]): Seq[Any] = {
-    // Collect URL query parameters and other parameters embedded inside URL.
-    val requestParams = request.query ++ extractPathParams(request)
-
-    // Build the function arguments
-    val methodArgs: Seq[Any] =
-      for (arg <- methodSurface.args) yield {
-        arg.surface.rawType match {
-          case cl if classOf[HttpRequest[_]].isAssignableFrom(cl) =>
-            // Bind the current http request instance
-            request
-          case _ =>
-            // Build from the string value in the request params
-            val argCodec = MessageCodecFactory.defaultFactory.of(arg.surface)
-            val v: Option[Any] = requestParams.get(arg.name) match {
-              case Some(paramValue) =>
-                // Pass the String parameter to the method argument
-                argCodec.unpackMsgPack(StringCodec.toMsgPack(paramValue))
-              case None =>
-                // Build the parameter from the content body
-                val contentBytes = request.contentBytes
-
-                if (contentBytes.nonEmpty) {
-                  val msgpack =
-                    request.contentType.map(_.split(";")(0)) match {
-                      case Some("application/x-msgpack") =>
-                        contentBytes
-                      case Some("application/json") =>
-                        // JSON -> msgpack
-                        JSONCodec.toMsgPack(JSON.parse(contentBytes))
-                      case _ =>
-                        // Try parsing as JSON first
-                        Try(JSON.parse(contentBytes))
-                          .map { jsonValue =>
-                            JSONCodec.toMsgPack(jsonValue)
-                          }
-                          .getOrElse {
-                            // If parsing as JSON fails, treat the content body as a regular string
-                            StringCodec.toMsgPack(request.contentString)
-                          }
-                    }
-                  argCodec.unpackMsgPack(msgpack)
-                } else {
-                  // Return the method default argument if exists
-                  arg.getMethodArgDefaultValue(controller)
-                }
-            }
-            // If mapping fails, use the zero value
-            v.getOrElse(Zero.zeroOf(arg.surface))
-        }
+  private[http] def checkPath(requestPathComponents: Seq[String], routePathComponents: Seq[String]): Boolean = {
+    if (requestPathComponents.length == routePathComponents.length) {
+      requestPathComponents.zip(routePathComponents).forall {
+        case (requestPathComponent, routePathComponent) =>
+          routePathComponent.startsWith(":") || routePathComponent == requestPathComponent
       }
-    trace(s"(${methodSurface.args.mkString(", ")}) <=  [${methodArgs.mkString(", ")}]")
-    methodArgs
-  }
-
-  def call(controller: Any, methodArgs: Seq[Any]): Any = {
-    methodSurface.call(controller, methodArgs: _*)
-  }
-
-  def call[A](controllerProvider: ControllerProvider, request: HttpRequest[A]): Option[Any] = {
-    controllerProvider.findController(controllerSurface).map { controller =>
-      call(controller, buildControllerMethodArgs(controller, request))
+    } else {
+      false
     }
   }
 }
